Performance upgrades. New Profiler class. Collision test scene

+++ main/include/Texture.hpp
@@ -16,10 +16,8 @@ class Texture {
     int getWidth() const { return width; }
     int getHeight() const { return height; }
 
-    Color sample(float u, float v) const;
+    Color sample(int u, int v) const;
     void setWrapMode(const char *mode) { wrapMode = mode; }
 
-    // Static factory method - you'll need to implement BMP loading based on
-    // your needs
     static Texture fromBMP(const uint8_t *data, size_t length);
 };
