+++ main/src/Shapes/Point.cpp
@@ -6,10 +6,10 @@ Collider *Point::defaultCollider() { return new PointCollider(x, y); }
 
 Pixels Point::drawAntiAliased() {
     Color sampledColor = sampleTexture(x, y);
-    return {Pixel(static_cast<int>(x), static_cast<int>(y), sampledColor)};
+    return {Pixel(x, y, sampledColor)};
 }
 
 Pixels Point::drawAliased() {
     Color sampledColor = sampleTexture(x, y);
-    return {Pixel(static_cast<int>(x), static_cast<int>(y), sampledColor)};
+    return {Pixel(x, y, sampledColor)};
 }
